requirements.txt
streamlit==1.31.0
pandas
plotly.express
requests
xlxswriter
lxml
